tutorial/urls.py
from django.urls import include, path
from rest_framework import routers
from quickstart import views
from rest_framework.schemas import get_schema_view


schema_view = get_schema_view(title='Pastebin API')

router = routers.DefaultRouter()
router.register(r'users', views.UserViewSet)
router.register(r'groups', views.GroupViewSet)

# Wire up our API using automatic URL routing.
# Additionally, we include login URLs for the browsable API.
urlpatterns = [
    path('schema/', schema_view),
    path('', include(router.urls)),
    path('', include('snippets.urls')),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework'))
]
